+++ modules/public_pages/erp/manufacturing/models/STItem.php
@@ -94,12 +94,13 @@ class STItem extends DataObject
 		// set formatters, more set in load() function
 
 		// Define enumerated types
-		$this->setEnum('comp_class',array( 'B'=>'Bought In'
-										  ,'M'=>'Manufactured'
-										  ,'S'=>'Sub-Contracted'));
-		$this->setEnum('abc_class',array( 'A'=>'A'
-                                         ,'B'=>'B'
-                                         ,'C'=>'C'));
+		$this->setEnum('comp_class', array('B' => 'Bought In',
+										   'K' => 'Sales Kit',
+										   'M' => 'Manufactured',
+										   'S' => 'Sub-Contracted'));
+		$this->setEnum('abc_class', array( 'A' => 'A',
+                                           'B'=>'B',
+                                           'C'=>'C'));
 		$this->setEnum('cost_basis', [
 		    'VOLUME' => 'Volume',
 		    'TIME' => 'Time'
@@ -140,6 +141,7 @@ class STItem extends DataObject
 											 ,'modules'=>array('new'=>array('module'=>'purchase_order'))
 											 ,'actions'=>array('link','new')
 											 ,'rules'=>array(array('field'=>'comp_class', 'criteria'=>"!='M'")
+											 				,array('field'=>'comp_class', 'criteria'=>"!='K'")
 											 				,array('field'=>'relatedCount("po_products")', 'criteria'=>'>0'))
 											 )
    						,'so_product_prices'=>array('newtab'=>array('new'=>true)
@@ -151,6 +153,7 @@ class STItem extends DataObject
 											 ,'modules'=>array('new'=>array('module'=>'purchase_order'))
 											 ,'actions'=>array('new')
 											 ,'rules'=>array(array('field'=>'comp_class', 'criteria'=>"!='M'")
+											 				,array('field'=>'comp_class', 'criteria'=>"!='K'")
 											 				,array('field'=>'relatedCount("po_products")', 'criteria'=>'==0'))
 											 )
    						,'so_products'=>array('newtab'=>array('new'=>true)
@@ -334,6 +337,9 @@ class STItem extends DataObject
 	public function calcLatestCost()
 	{
 		switch ($this->comp_class) {
+			case 'K':
+				$var[] = $this->calcLatestMat();
+				break;
 			case 'M':
 			case 'S':
 				$var[] = $this->calcLatestMat();